Create fibonacci
Fibonacci practice.

First, I defined a function "fibo(n)." Then I transformed the type of "n" as an integer. (If I don't do it, Python will consider it as "string" - which will prevent "n" from being added or subtracted)

Then I went with three variables. In this case, I believe we will need three variables. The first one is to "show" the answer for the given n. The second one is to be added with the first one so that the series can make a progress. The third one is to memorize what happened before.

This one is quite an easy one, but it took like 30 minutes for me to figure out. But I am totally fine with that! All experts were once beginners!