Create mine_finder
* denotes a cell with a mine; . denotes a cell without a mine.
This code is to find the number of mines adjacent to a certain cell. If the cell contains a mine, then the output should be "*".

For example:
input:
3 3
.**
*..
.*.

output:
2**
*43
2*1

While there is a lot of room to improve for this code, but I am more than happy with solving this problem on my own!